Separate `fake' descriptor sets from the GL implementation
